Avoid rendering fills if empty elements inside (#20495)
